Complete refactor of `check_int` - into a better location
