+++ portal/type_tools.py
@@ -1,3 +1,5 @@
+from __future__ import unicode_literals  # isort:skip
+
 from flask import abort
